Update main.css
Slightly updated index.html and main.css after merging from main. Changed the styling for the body tag to reference the main tag in main.css. Added a main tag to index.html. This was to allow the header to span the whole page. We are changing the main branch to be more grid based so this commit may not be necessary in the long run.