ChangeListener and Booking
Added a remove function to the ChangeListener class so that we can stop listening to files that we are no longer accessing.

Updated booking.js to fix the issue where the latest seat selection was not being saved to the booking JSON.

Updated booking.js so that the changeListener works on the file being accessed.

Updated the updateBookingJSON function so that we only update the checkbox (seat) that was clicked and not the entire seating array.